- Added command line extract option to write streams out to files. - Created a StreamType unit to manage stream types.

+++ inc/PacketParser.h
@@ -31,7 +31,7 @@ namespace TS
         void parse_PAT_table(PacketBuffer& p_buffer);
         void parse_PMT_table(PacketBuffer& p_buffer);
         void parse_elementary_stream_specific_data(PacketBuffer& p_buffer, uint16_t elementary_stream_specific_data_size);
-        std::vector<Descriptor> parse_descriptors(PacketBuffer& p_buffer, uint8_t descriptors_size);
+        std::vector<Descriptor> parse_descriptors(PacketBuffer& p_buffer, uint16_t descriptors_size);
 
         Packet _packet{};
     };
